feat(app): refactor copy to clipboard as effect
